refactor: remove streaming_v2 module and update streaming functionality
- Deleted the streaming_v2 module to streamline the codebase and consolidate streaming logic.
- Updated the StreamingRenderer and BackgroundPromiseResolver to utilize the existing streaming module.
- Refactored related components to improve error handling and logging during streaming processes.
- Adjusted API routes to reflect the removal of the legacy streaming endpoint.
